packages/htn/packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.3.0",
    "pandas>=2.3.2",
    "python-multipart>=0.0.20",
    "scalar-fastapi>=1.4.0",
]
